+++ app/resources/views/pages/dominion/op-center/index.blade.php
@@ -8,7 +8,7 @@
         <div class="col-sm-12 col-md-9">
             <div class="box box-primary">
                 <div class="box-header">
-                    <h3 class="box-title"><i class="fa fa-globe"></i> Op Center</h3>
+                    <h3 class="box-title"><i class="ra ra-scroll-unfurled"></i> Op Center</h3>
                 </div>
                 <div class="box-body table-responsive">
                     <table class="table table-hover" id="dominions-table">