+++ app/resources/views/partials/main-sidebar.blade.php
@@ -128,7 +128,7 @@
                     </a>
                 </li>
                 <li class="{{ Route::is('dominion.espionage') ? 'active' : null }}"><a href="{{ route('dominion.espionage') }}"><i class="fa fa-user-secret fa-fw"></i> <span>Espionage</span></a></li>
-                <li class="{{ Route::is('dominion.op-center*') ? 'active' : null }}"><a href="{{ route('dominion.op-center') }}"><i class="fa fa-globe fa-fw"></i> <span>Op Center</span></a></li>
+                <li class="{{ Route::is('dominion.op-center*') ? 'active' : null }}"><a href="{{ route('dominion.op-center') }}"><i class="ra ra-scroll-unfurled ra-fw"></i> <span>Op Center</span></a></li>
                 <li class="{{ Route::is('dominion.bounty-board') ? 'active' : null }}">
                     <a href="{{ route('dominion.bounty-board') }}">
                         <i class="ra ra-hanging-sign ra-fw"></i> <span>Bounty Board</span>
@@ -152,6 +152,7 @@
 
                 <li class="header">RELATIONS</li>
                 <li class="{{ Route::is('dominion.realm') ? 'active' : null }}"><a href="{{ route('dominion.realm') }}"><i class="ra ra-circle-of-circles ra-fw"></i> <span>Realms</span></a></li>
+                <li class="{{ Route::is('dominion.world') ? 'active' : null }}"><a href="{{ route('dominion.world') }}"><i class="fa fa-globe fa-fw"></i> <span>The World</span></a></li>
                 <li class="{{ Route::is('dominion.search') ? 'active' : null }}"><a href="{{ route('dominion.search') }}"><i class="fa fa-search fa-fw"></i> <span>Search</span></a></li>
                 <li class="{{ Route::is('dominion.council*') ? 'active' : null }}">
                     <a href="{{ route('dominion.council') }}">